load cart item

+++ app/src/main/java/com/tofukma/orderapp/Database/LocalCartDataSource.kt
@@ -22,7 +22,7 @@ class LocalCartDataSource(private val cartDAO: CartDAO) : CartDataSource {
         return cartDAO.countItemInCart(uid)
     }
 
-    override fun sumPrice(uid: String): Single<Int> {
+    override fun sumPrice(uid: String): Single<Double> {
         return cartDAO.sumPrice(uid)
     }
